feat: add status component
